chore: New 'types' module
Migrate Request types for run.create methods to new types module
