+++ src/leapfrogai_api/types/constants.py
@@ -1,0 +1,6 @@
+# This file defines application constants not set by environment variables or configuration files.
+
+THIRTY_DAYS = 60 * 60 * 24 * 30  # in seconds
+
+DEFAULT_MAX_COMPLETION_TOKENS = 4096
+DEFAULT_MAX_PROMPT_TOKENS = 4096
