+++ src/leapfrogai_api/types/model_types.py
@@ -1,0 +1,10 @@
+from typing import List
+
+
+class Model:
+    name: str
+    backend: str
+
+    def __init__(self, name: str, backend: str, capabilities: List[str] | None = None):
+        self.name = name
+        self.backend = backend
